backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
openai==1.51.0
spacy==3.7.6
PyMuPDF==1.24.10
python-docx==1.1.2
sentence-transformers==3.1.1
scikit-learn==1.5.2
numpy==1.26.4
shap==0.46.0
dice-ml==0.11
aiofiles==24.1.0
jinja2==3.1.4
